move taker restart functionality to a separate module
